--- instrumentation/redisson/redisson-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/redisson/RedissonNetAttributesGetter.java
@@ -1,19 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.redisson;
-
-import io.opentelemetry.instrumentation.api.semconv.network.NetworkAttributesGetter;
-import java.net.InetSocketAddress;
-import javax.annotation.Nullable;
-
-final class RedissonNetAttributesGetter implements NetworkAttributesGetter<RedissonRequest, Void> {
-
-  @Override
-  public InetSocketAddress getNetworkPeerInetSocketAddress(
-      RedissonRequest request, @Nullable Void unused) {
-    return request.getAddress();
-  }
-}
